Try to quickly detect whether yarn install is needed at startup.
